+++ Application/module/person/callbacks.cpp
@@ -12,8 +12,8 @@ void ModulePerson::Callbacks()
 v8::Handle<v8::Value> ModulePerson::jsPlayer(const v8::Arguments& args)
 {
 	ModulePerson *module = (ModulePerson*)HelperScript::Unwrap(args.Data());
-	auto cam = module->Entity->Get<Camera>(*module->Global->Get<unsigned int>("camera"));
-	unsigned int id = cam->Person;
+	auto cam = module->Entity->Get<Camera>(*module->Global->Get<uint64_t>("camera"));
+	uint64_t id = cam->Person;
 
 	return v8::Uint32::New(id);
 }
